+++ pii-secret-exclude.txt
@@ -148,3 +148,4 @@ international/templates/international/base.html
 international/migrations/0001_initial.py
 tests/unit/sso_profile/business_profile/test_helpers.py
 tests/unit/core/management/commands/test_send_campaign_site_review_reminder.py
+tests/unit/international/test_forms.py
