+++ tests/unit/international_online_offer/test_forms.py
@@ -2,7 +2,6 @@ import pytest
 
 from international_online_offer.core import hirings, intents, regions, spends
 from international_online_offer.forms import (
-    ContactForm,
     CsatFeedbackForm,
     FeedbackForm,
     HiringForm,
@@ -224,40 +223,4 @@ def test_feedback_form_validation(form_data, is_valid):
 def test_csat_feedback_form_validation(form_data, is_valid):
     data = form_data
     form = CsatFeedbackForm(data)
-    assert form.is_valid() == is_valid
-
-
-@pytest.mark.parametrize(
-    'form_data,is_valid',
-    (
-        (
-            {
-                'full_name': 'Jane Bloggs',
-                'email': 'test@test.com',
-                'how_we_can_help': 'Please help me login',
-            },
-            True,
-        ),
-        (
-            {
-                'full_name': '',
-                'email': '',
-                'how_we_can_help': '',
-            },
-            False,
-        ),
-        (
-            {
-                'full_name': 'Joe Bloggs',
-                'email': 'incorrect email',
-                'how_we_can_help': 'Please help me login',
-            },
-            False,
-        ),
-    ),
-)
-@pytest.mark.django_db
-def test_contact_form_validation(form_data, is_valid):
-    data = form_data
-    form = ContactForm(data)
     assert form.is_valid() == is_valid
